remove old state trans. components
